Results Tab: The 'Save' button now saves the correct results page. The first results page has the file name entered correctly now as well.
These two issues were mainly due to the code assuming a base of 1 for tab indexes when it is actually 0.

I also simply added the HTML as a property of the actual Result QWebView objects instead of trying to keep a separately synched array of results. This should make the code easier to modify and allow more flexibility for future changes. E.g. re-ordering or sorting of tabs. The code for the arrays is still present - just commented out (4 lines).