Barrage: Modified mechanics to replicate tested in-game behavior. See below:
Buff Refresh: The buff can now be refreshed as soon as the 6 sec icd is up instead of waiting the entire 15 seconds for it to expire (or to be consumed). Verified in game. Also added a hit check for Tracer Missile.

Buff Consumption: Moved to the last offhand tick since it still triggers on a miss. 'TODO:' comment removed.

Unload Cooldown Reset: The instant cooldown reset caused Unload to be used instantly on triggering which is impossible in-game. I've overloaded cooldown.reset() to allow specifying an override time. That amount of time will be left on the cooldown unless it is already lower (time will never be added) to model player reaction time. This seemed like a much simpler and more appropriate fix than trying to delay the cd reset. Note: I've used sim->reaction_time instead of p.total_reaction_time() since it sometimes produces values of over 2 seconds! As a side note, a default value of 0.25 for reaction_time (instead of 0.5) may be more appropriate in general according to this: http://www.humanbenchmark.com/tests/reactiontime/stats.php

From the documentation cooldown_react looked like it would be ideal for this but doesn't seem to be implemented for actions?

Unload Tick Benefit: Changed so every Unload tick checks for barrage instead of a single check on execute. This reflects tested in-game behavior. If the buff expires half way through Unload, subsequent ticks don't benefit.

Unload Offhand Channel: Anything that interrupted the MH channel would still leave the OH channel to tick for the full duration. I've set channeled=true for the OH action to better reflect the actual behavior. A more complete fix would be ideal. There is a detailed comment in the code.

Power Shot: Only procs Barrage off the main hand hit so the 'TEST:' comment has been removed. Tested with over 1000 samples in-game waiting for 6 sec cd & considering offhand misses.