Added Dread Forged and Obroan set gear and relics.
Added a wrapper function for get_set_bonus that simplified the arguements a lot.

It now just takes a list of "/" seperated gear prefixes and a "/" seperated list of set name suffixes and generates a string containing all combinations of the two.

The set name is created by combimimg the first entry of arguement.